refactor: drop the pkg dir
BREAKING CHANGE: rename packages
