+++ defaults_command.go
@@ -12,7 +12,7 @@ import (
 	"sort"
 
 	"github.com/MakeNowJust/heredoc/v2"
-	"github.com/nais/naistrix/pkg/input"
+	"github.com/nais/naistrix/input"
 	"github.com/spf13/viper"
 )
