+++ output/link.go
